Renamed StatusError to PodInfo
